+++ tests/auxtel/test_build_pointing_model.py
@@ -18,20 +18,24 @@
 #
 # You should have received a copy of the GNU General Public License
 
-import types
-import unittest
 import contextlib
-import unittest.mock
 import os
 import tempfile
+import types
+import unittest
+import unittest.mock
 
 import numpy as np
-
-from lsst.ts.standardscripts import BaseScriptTestCase
+from lsst.daf import butler as dafButler
+from lsst.ts import salobj
 from lsst.ts.externalscripts import get_scripts_dir
-from lsst.ts.externalscripts.auxtel.build_pointing_model import BuildPointingModel
+from lsst.ts.externalscripts.auxtel.build_pointing_model import (
+    BuildPointingModel,
+    GridType,
+    generate_rotator_sequence,
+)
 from lsst.ts.observatory.control.utils.enums import RotType
-import lsst.daf.butler as dafButler
+from lsst.ts.standardscripts import BaseScriptTestCase
 from lsst.utils import getPackageDir
 
 # Declare the local path that has the information to build a
@@ -64,14 +68,10 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
 
         return (self.script,)
 
-    async def test_configure(self):
-
-        async with self.make_configured_dry_script() as test_configuration:
-            self.assert_config(test_configuration)
-
     def assert_config(self, configuration):
         self.check_configuration_values(configuration)
         self.check_azel_grid()
+        self.check_rotator_gen()
 
     def check_configuration_values(self, configuration):
 
@@ -93,9 +93,96 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
             np.max(self.script.elevation_grid) <= self.script.config.elevation_maximum
         )
 
+        if self.script.config.grid == GridType.RADEC.name:
+            # We expect at least one point in the grid is rejected because of
+            # high elevation.
+            # This is why we subtract one from the number of points in the grid
+            # below.
+            expected_size = np.sum(self.script.config.radec_grid["ha_grid"]["n"]) - 1
+            assert len(self.script.elevation_grid) == expected_size
+
+    def check_rotator_gen(self):
+
+        expected_next_rot_seq = self.script.config.rotator_sequence
+
+        self._check_rotator_sequence(expected_next_rot_seq, reverse=False)
+
+        expected_next_rot_seq = self.script.config.rotator_sequence[:-1:][::-1]
+
+        self._check_rotator_sequence(expected_next_rot_seq, reverse=True)
+
+        expected_next_rot_seq = self.script.config.rotator_sequence[1::]
+
+        self._check_rotator_sequence(expected_next_rot_seq, reverse=False)
+
+    def _check_rotator_sequence(self, expected_next_rot_seq, reverse):
+
+        for rot_grid_expected, rot_grid_gen in zip(
+            expected_next_rot_seq, self.script.rotator_sequence_gen
+        ):
+
+            with self.subTest(rot_grid_expected=rot_grid_expected):
+
+                for rot_expected, rot_gen in zip(
+                    rot_grid_expected if not reverse else rot_grid_expected[::-1],
+                    rot_grid_gen,
+                ):
+                    assert rot_expected == rot_gen
+
+    def assert_metadata(self, metadata):
+
+        assert metadata.nimages == len(self.script.elevation_grid) * 2
+        assert metadata.duration == metadata.nimages * (
+            self.script.config.exposure_time
+            + self.script.camera_readout_time
+            + self.script.estimated_average_slew_time
+        )
+
+    async def test_configure_use_healpix_grid(self):
+
+        async with self.make_configured_dry_script(
+            grid=GridType.HEALPIX
+        ) as test_configuration:
+            self.assert_config(test_configuration)
+
+    async def test_configure_use_radec_grid(self):
+
+        async with self.make_configured_dry_script(
+            grid=GridType.RADEC
+        ) as test_configuration:
+            self.assert_config(test_configuration)
+
+    async def test_configure_skip(self):
+        async with self.make_configured_dry_script(
+            grid=GridType.RADEC,
+            skip=10,
+        ) as test_configuration:
+            self.assert_config(test_configuration)
+
+    async def test_configure_fails(self):
+
+        self.remotes_needed = False
+        async with self.make_script():
+            bad_config = [
+                dict(grid="this_is_a_typo"),
+                dict(healpix_grid=dict(nside=0)),
+                dict(radec_grid=dict(dec_grid=dict(min=-100))),
+                dict(radec_grid=dict(dec_grid=dict(max=100))),
+                dict(radec_grid=dict(dec_grid=dict(n=1))),
+                dict(radec_grid=dict(ha_grid=dict(min=-14))),
+                dict(radec_grid=dict(ha_grid=dict(max=14))),
+                dict(radec_grid=dict(ha_grid=dict(n=[]))),
+                dict(grid="radec", radec_grid=dict(ha_grid=dict(n=[3, 5, 6]))),
+            ]
+            for config in bad_config:
+                with self.subTest(config=config), self.assertRaises(
+                    salobj.ExpectedError
+                ):
+                    await self.configure_script(**config)
+
     async def test_metadata(self):
 
-        async with self.make_configured_dry_script():
+        async with self.make_configured_dry_script(grid=GridType.HEALPIX):
 
             metadata = types.SimpleNamespace(
                 duration=0.0,
@@ -107,15 +194,6 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
 
             self.assert_metadata(metadata)
 
-    def assert_metadata(self, metadata):
-
-        assert metadata.nimages == len(self.script.elevation_grid) * 2
-        assert metadata.duration == metadata.nimages * (
-            self.script.config.exposure_time
-            + self.script.camera_readout_time
-            + self.script.estimated_average_slew_time
-        )
-
     async def test_executable(self):
 
         scripts_dir = get_scripts_dir()
@@ -124,7 +202,17 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
 
     async def test_arun(self):
 
-        async with self.make_configured_dry_script():
+        async with self.make_configured_dry_script(grid=GridType.HEALPIX):
+
+            self.script.execute_grid = unittest.mock.AsyncMock()
+
+            await self.script.arun()
+
+            self.assert_arun()
+
+    async def test_arun_skip_some(self):
+
+        async with self.make_configured_dry_script(grid=GridType.HEALPIX, skip=10):
 
             self.script.execute_grid = unittest.mock.AsyncMock()
 
@@ -134,14 +222,33 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
 
     def assert_arun(self):
 
-        assert self.script.execute_grid.await_count == self.script.grid_size
-        execute_grid_calls = [
-            unittest.mock.call(azimuth, elevation)
-            for azimuth, elevation in zip(
-                self.script.azimuth_grid, self.script.elevation_grid
-            )
-        ]
-        self.script.execute_grid.assert_has_awaits(execute_grid_calls)
+        self.script.latiss.rem.atspectrograph.cmd_changeFilter.set_start.assert_awaited_with(
+            filter=0, name="empty_1", timeout=self.script.latiss.long_timeout
+        )
+        self.script.latiss.rem.atspectrograph.cmd_changeDisperser.set_start.assert_awaited_with(
+            disperser=0, name="empty_1", timeout=self.script.latiss.long_timeout
+        )
+
+        rot_seq_gen = generate_rotator_sequence(self.script.config.rotator_sequence)
+
+        expected_execute_grid_calls = []
+
+        grid_index = range(len(self.script.azimuth_grid))
+        for index, azimuth, elevation, rot_seq in zip(
+            grid_index,
+            self.script.azimuth_grid,
+            self.script.elevation_grid,
+            rot_seq_gen,
+        ):
+            if index < self.script.config.skip:
+                continue
+            for rot in rot_seq:
+                expected_execute_grid_calls.append(
+                    unittest.mock.call(azimuth, elevation, rot)
+                )
+
+        assert self.script.execute_grid.await_count == len(expected_execute_grid_calls)
+        self.script.execute_grid.assert_has_awaits(expected_execute_grid_calls)
 
         ataos_reset_offset_calls = [
             unittest.mock.call(axis="x", timeout=self.script.atcs.long_timeout),
@@ -170,11 +277,12 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
 
     async def test_execute_grid(self):
 
-        async with self.make_configured_dry_script():
+        async with self.make_configured_dry_script(grid=GridType.HEALPIX):
 
-            azimuth, elevation = (
+            azimuth, elevation, rotator = (
                 self.script.azimuth_grid[0],
                 self.script.elevation_grid[0],
+                np.random.rand(),
             )
 
             target_name = "HD 12345"
@@ -185,13 +293,18 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
             self.script.atcs.slew_object = unittest.mock.AsyncMock()
             self.script.center_on_brightest_source = unittest.mock.AsyncMock()
 
-            await self.script.execute_grid(azimuth=azimuth, elevation=elevation)
-
-            self.assert_execute_grid(
-                azimuth=azimuth, elevation=elevation, target_name=target_name
+            await self.script.execute_grid(
+                azimuth=azimuth, elevation=elevation, rotator=rotator
             )
 
-    def assert_execute_grid(self, azimuth, elevation, target_name):
+            self.assert_execute_grid(
+                azimuth=azimuth,
+                elevation=elevation,
+                rotator=rotator,
+                target_name=target_name,
+            )
+
+    def assert_execute_grid(self, azimuth, elevation, rotator, target_name):
 
         self.script.atcs.find_target.assert_awaited_once_with(
             az=azimuth,
@@ -201,6 +314,7 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
         )
         self.script.atcs.slew_object.assert_awaited_once_with(
             name=target_name,
+            rot=rotator,
             rot_type=RotType.PhysicalSky,
         )
         self.script.center_on_brightest_source.assert_awaited()
@@ -209,7 +323,7 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
 
     async def test_execute_grid_fail_to_find_target(self):
 
-        async with self.make_configured_dry_script():
+        async with self.make_configured_dry_script(grid=GridType.HEALPIX):
 
             azimuth, elevation = (
                 self.script.azimuth_grid[0],
@@ -222,7 +336,9 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
             self.script.atcs.slew_object = unittest.mock.AsyncMock()
             self.script.center_on_brightest_source = unittest.mock.AsyncMock()
 
-            await self.script.execute_grid(azimuth=azimuth, elevation=elevation)
+            await self.script.execute_grid(
+                azimuth=azimuth, elevation=elevation, rotator=0.0
+            )
 
             self.assert_execute_grid_fail(azimuth=azimuth, elevation=elevation)
 
@@ -241,7 +357,7 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
 
     async def test_center_on_brightest_source(self):
 
-        async with self.make_configured_dry_script():
+        async with self.make_configured_dry_script(grid=GridType.HEALPIX):
 
             self.find_offset_image_id = [
                 123,
@@ -305,34 +421,57 @@ class TestBuildPointingModel(BaseScriptTestCase, unittest.IsolatedAsyncioTestCas
 
         self.script.atcs.add_point_data.assert_awaited_once()
 
-    async def set_test_configuration(self):
+    async def set_test_configuration(self, grid, **kwargs):
 
-        test_configuration = dict(
-            nside=3,
-            azimuth_origin=-265,
-            elevation_minimum=20.0,
-            elevation_maximum=80.0,
-            magnitude_limit=9.0,
-            exposure_time=1.0,
-        )
+        test_configuration = self._generate_configuration(grid=grid, **kwargs)
 
         await self.configure_script(**test_configuration)
 
         return test_configuration
 
+    def _generate_configuration(self, grid, **kwargs):
+        """Generate configuration.
+
+        Parameters
+        ----------
+        grid : `str`
+            Grid name,
+
+        Returns
+        -------
+        configuration : `dict`
+            Script configuration.
+        """
+        configuration = dict(
+            grid=grid.name.lower(),
+            elevation_minimum=20.0,
+            elevation_maximum=80.0,
+            magnitude_limit=9.0,
+            exposure_time=1.0,
+        )
+        configuration.update(**kwargs)
+
+        return configuration
+
     @contextlib.asynccontextmanager
-    async def make_configured_dry_script(self):
-        """Construct script without remotes.
+    async def make_configured_dry_script(self, grid, **kwargs):
+        """Construct script without remotes with healpix grid configuration.
 
         This is useful for developing fast unit tests for methods that do not
         require DDS communication or when mocking the remote's behavior.
-        """
 
+        Parameters
+        ----------
+        grid : `str
+            Which grid to configure the the script for.
+        """
         self.remotes_needed = False
         async with self.make_script():
 
-            test_configuration = await self.set_test_configuration()
+            test_configuration = await self.set_test_configuration(grid=grid, **kwargs)
+
             self.script.atcs.rem.ataos = unittest.mock.AsyncMock()
             self.script.atcs.rem.atptg = unittest.mock.AsyncMock()
+            self.script.latiss.rem.atspectrograph = unittest.mock.AsyncMock()
 
             yield test_configuration